refactor: widgets externalization
  - Logo widget separated from login_screen
  - Widgets export file created
